+++ src/main/java/store/lastdance/service/expense/ExpenseV2ServiceImpl.java
@@ -4,9 +4,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -16,7 +13,6 @@ import store.lastdance.domain.group.Group;
 import store.lastdance.domain.group.GroupMember;
 import store.lastdance.domain.user.User;
 import store.lastdance.dto.expense.*;
-import store.lastdance.dto.response.PageWithSummaryResponse;
 import store.lastdance.exception.CustomException;
 import store.lastdance.exception.ErrorCode;
 import store.lastdance.repository.expense.ExpenseAnalysisHistoryRepository;
@@ -30,16 +26,14 @@ import store.lastdance.service.image.ImageService;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.TemporalAdjusters;
 import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
+@Transactional
 @Slf4j
-public class ExpenseV2ServiceImpl implements ExpenseService {
+public class ExpenseV2ServiceImpl implements ExpenseV2Service {
     private final ExpenseRepository expenseRepository;
     private final ExpenseSplitRepository expenseSplitRepository;
     private final GroupRepository groupRepository;
@@ -66,7 +60,6 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
      * 개인 지출 등록
      */
     @Override
-    @Transactional
     public ExpenseResponseDTO createPersonalExpense(UUID userId, CreatePersonalExpenseRequestDTO requestDTO, MultipartFile receiptFile) {
         Expense expense = createBaseExpense(userId, requestDTO, ExpenseType.PERSONAL, receiptFile);
 
@@ -78,7 +71,6 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
      * 그룹 지출 등록
      */
     @Override
-    @Transactional
     public ExpenseResponseDTO createGroupExpense(UUID userId, CreateGroupExpenseRequestDTO requestDTO, MultipartFile receiptFile) {
         Expense expense = createBaseExpense(userId, requestDTO, ExpenseType.GROUP, receiptFile);
 
@@ -251,31 +243,11 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
                 .toList();
     }
 
-    /**
-     * 지출 조회
-     */
-    @Override
-    public ExpenseResponseDTO getExpenseById(UUID userId, Long expenseId) {
-        User user = findUserById(userId);
-
-        Expense expense = expenseRepository.findByExpenseIdWithPermission(expenseId, user).orElseThrow(
-                () -> new CustomException(ErrorCode.EXPENSE_NOT_FOUND)
-        );
-
-        // 그룹 지출 - 정산 데이터 포함
-        List<SplitDataDTO> splitData = null;
-        if (expense.getExpenseType() == ExpenseType.GROUP) {
-            splitData = getSplitData(expense);
-        }
-
-        return ExpenseResponseDTO.from(expense, splitData);
-    }
 
     /**
      * 지출 수정
      */
     @Override
-    @Transactional
     public ExpenseResponseDTO updateExpense(UUID userId, Long expenseId, UpdateExpenseRequestDTO requestDTO, MultipartFile receiptFile) {
 
         User user = findUserById(userId);
@@ -364,7 +336,6 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
      * 지출 삭제
      */
     @Override
-    @Transactional
     public void deleteExpense(UUID userId, Long expenseId) {
 
         User user = findUserById(userId);
@@ -382,84 +353,11 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
         expenseRepository.deleteById(expenseId);
     }
 
-    /**
-     * 그룹 분담금 조회
-     */
-    @Override
-    public List<GroupShareExpenseResponseDTO> getGroupShareExpenses(UUID userId, ExpenseSearchDTO searchDTO) {
-        log.info("=== getGroupShareExpenses 호출 ===");
-        log.info("userId: {}, year: {}, month: {}", userId, searchDTO.year(), searchDTO.month());
-
-        User user = findUserById(userId);
-
-        // 1. SHARE 타입 지출들 조회
-        List<Expense> shareExpenses = expenseRepository.findShareExpensesByUserAndMonth(user, searchDTO.year(), searchDTO.month());
-        log.info("조회된 SHARE 지출 개수: {}", shareExpenses.size());
-
-        return shareExpenses.stream()
-                .map(shareExpense -> {
-                    log.debug("--- SHARE 지출 처리 중 ---");
-                    log.debug("SHARE 지출 ID: {}", shareExpense.getExpenseId());
-                    log.debug("SHARE 지출 제목: {}", shareExpense.getTitle());
-                    log.debug("SHARE 분담 금액: {}", shareExpense.getAmount());
-                    log.debug("원본 지출 ID: {}", shareExpense.getOriginalExpense() != null ? shareExpense.getOriginalExpense().getExpenseId() : null);
-
-                    // 2. 원본 그룹 지출 조회 (더 많은 정보를 위해)
-                    Expense originalExpense = shareExpense.getOriginalExpense();
-
-                    // 3. 분할 정보 조회 (원본 지출 기준)
-                    List<SplitDataDTO> splitData = null;
-                    if (originalExpense != null) {
-                        splitData = getSplitData(originalExpense);
-                    }
-
-                    // 4. 그룹 이름 조회
-                    String groupName = shareExpense.getGroup() != null ?
-                            shareExpense.getGroup().getGroupName() : "";
-                    log.debug("그룹 이름: {}", groupName);
-
-
-                    GroupShareExpenseResponseDTO result = GroupShareExpenseResponseDTO.from(
-                            shareExpense,
-                            originalExpense,
-                            groupName,
-                            splitData
-                    );
-                    log.debug("생성된 응답 데이터 - expenseId: {}, title: {}, myShareAmount: {}",
-                            result.expenseId(), result.title(), result.myShareAmount());
-
-                    return result;
-
-                })
-                .toList();
-    }
-
-    /**
-     * 영수증 조회
-     */
-    @Override
-    public String getReceiptImageUrl(Long expenseId, UUID userId) {
-
-        User user = findUserById(userId);
-        // 권한 체크 포함해서 지출 조회
-        Expense expense = expenseRepository.findByExpenseIdWithPermission(expenseId, user).orElseThrow(
-                () -> new CustomException(ErrorCode.EXPENSE_NOT_FOUND)
-        );
-
-        // 영수증 파일 없으면 null 반환
-        ImageFile receiptImageFile = expense.getReceiptImageFile();
-        if (receiptImageFile == null) {
-            return null;
-        }
-
-        return imageService.generatePresignedUrl(receiptImageFile.getFileId());
-    }
 
     /**
      * 영수증만 삭제 (지출은 유지)
      */
     @Override
-    @Transactional
     public void deleteReceiptImage(Long expenseId, UUID userId) {
 
         User user = findUserById(userId);
@@ -480,568 +378,15 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
         expense.updateReceiptImageFile(null);
     }
 
-    /**
-     * 개인 지출 월별 추이 조회
-     */
-    @Override
-    public MonthlyExpenseTrendResponseDTO getPersonalExpenseTrend(UUID userId, ExpenseSearchDTO searchDTO) {
-        log.info("개인 지출 추이 조회: userId={}, year={}, month={}, months={}, category={}", userId, searchDTO.year(), searchDTO.month(), searchDTO.months(), searchDTO.category());
-
-        User user = findUserById(userId);
-
-        // 날짜 범위 계산
-        DateRange dateRange = calculateDateRange(searchDTO.year(), searchDTO.month(), searchDTO.months());
-
-        ExpenseCategory categoryEnum = searchDTO.category() != null ? ExpenseCategory.valueOf(searchDTO.category().toUpperCase()) : null;
-        // Repository 조회
-        List<Expense> expenses = expenseRepository.findPersonalExpensesByMonthRange(user, dateRange.startDate(), dateRange.endDate(), categoryEnum);
-
-        // 월별 그룹핑 및 응답
-        return createTrendResponse(expenses, dateRange);
-    }
-
-    /**
-     * 그룹 지출 월별 추이 조회
-     */
-    @Override
-    public MonthlyExpenseTrendResponseDTO getGroupExpenseTrend(UUID userId, UUID groupId, ExpenseSearchDTO searchDTO) {
-        User user = findUserById(userId);
-        Group group = findGroupById(groupId);
-
-        // 그룹 멤버인지 확인
-        boolean isMember = groupMemberRepository.existsByGroupAndUser(group, user);
-        if (!isMember) {
-            throw new CustomException(ErrorCode.GROUP_MEMBER_NOT_FOUND);
-        }
-
-        log.info("그룹 지출 추이 조회: groupId={}, year={}, month={}, months={}, category={}",
-                groupId, searchDTO.year(), searchDTO.month(), searchDTO.months(), searchDTO.category());
-
-        // 날짜 범위 계산
-        DateRange dateRange = calculateDateRange(searchDTO.year(), searchDTO.month(), searchDTO.months());
-
-        ExpenseCategory categoryEnum = searchDTO.category() != null ? ExpenseCategory.valueOf(searchDTO.category().toUpperCase()) : null;
-        // Repository에서 데이터 조회
-        List<Expense> expenses = expenseRepository.findGroupExpensesByMonthRange(group, dateRange.startDate(), dateRange.endDate(), categoryEnum);
-
-        // 월별 그룹핑 및 응답 생성
-        return createTrendResponse(expenses, dateRange);
-    }
-
-    /**
-     * 날짜 범위 계산
-     */
-    private DateRange calculateDateRange(int year, int month, int months) {
-        LocalDate endDate = LocalDate.of(year, month, 1).with(TemporalAdjusters.lastDayOfMonth());
-        LocalDate startDate = endDate.minusMonths(months - 1).with(TemporalAdjusters.firstDayOfMonth());
-
-        return new DateRange(startDate, endDate);
-    }
-
-    /**
-     * 월별 추이 응답 생성
-     */
-    private MonthlyExpenseTrendResponseDTO createTrendResponse(List<Expense> expenses, DateRange dateRange) {
-        Map<String, List<ExpenseResponseDTO>> monthlyData = createMonthlyGrouping(expenses, dateRange);
-        return MonthlyExpenseTrendResponseDTO.create(monthlyData, dateRange.startDate, dateRange.endDate);
-    }
-
-    /**
-     * 지출 데이터를 월별로 그룹핑
-     */
-    private Map<String, List<ExpenseResponseDTO>> createMonthlyGrouping(List<Expense> expenses, DateRange dateRange) {
-        // N+1 문제 해결: 모든 분담 내역을 한 번에 조회
-        List<Expense> groupExpenses = expenses.stream()
-                .filter(e -> e.getExpenseType() == ExpenseType.GROUP)
-                .toList();
-
-        Map<Long, List<SplitDataDTO>> splitsByExpenseId = new HashMap<>();
-        if (!groupExpenses.isEmpty()) {
-            List<ExpenseSplit> allSplits = expenseSplitRepository.findByExpenseIn(groupExpenses);
-            splitsByExpenseId = allSplits.stream()
-                    .collect(Collectors.groupingBy(
-                            split -> split.getExpense().getExpenseId(),
-                            Collectors.mapping(
-                                    split -> new SplitDataDTO(split.getUser().getUserId(), split.getAmount()),
-                                    Collectors.toList()
-                            )
-                    ));
-        }
-
-        // 지출 데이터 월별로 그룹핑
-        final Map<Long, List<SplitDataDTO>> finalSplitsMap = splitsByExpenseId;
-        Map<String, List<ExpenseResponseDTO>> monthlyData = expenses.stream()
-                .collect(Collectors.groupingBy(
-                        expense -> expense.getExpenseDate().format(DateTimeFormatter.ofPattern("yyyy-MM")),
-                        LinkedHashMap::new,  // 순서 보장
-                        Collectors.mapping(
-                                expense -> ExpenseResponseDTO.from(expense, finalSplitsMap.get(expense.getExpenseId())),
-                                Collectors.toList()
-                        )
-                ));
-
-        // 데이터 없는 달 빈 배열로 추가
-        fillEmptyMonths(monthlyData, dateRange.startDate, dateRange.endDate);
-        return sortByMonthKey(monthlyData);
-    }
-
-    /**
-     * 데이터 업는 달에 빈 배열 추가
-     */
-    private void fillEmptyMonths(Map<String, List<ExpenseResponseDTO>> monthlyData, LocalDate startDate, LocalDate endDate) {
-        LocalDate current = startDate;
-        while (!current.isAfter(endDate)) {
-            String monthKey = current.format(DateTimeFormatter.ofPattern("yyyy-MM"));
-            monthlyData.putIfAbsent(monthKey, new ArrayList<>());
-            current = current.plusMonths(1);
-        }
-    }
-
-    /**
-     * 월별 키로 정렬
-     */
-    private Map<String, List<ExpenseResponseDTO>> sortByMonthKey(Map<String, List<ExpenseResponseDTO>> monthlyData) {
-        return monthlyData.entrySet().stream()
-                .sorted(Map.Entry.comparingByKey())
-                .collect(LinkedHashMap::new, (m, e) -> m.put(e.getKey(), e.getValue()),
-                        LinkedHashMap::putAll);
-    }
-
-    private record DateRange(LocalDate startDate, LocalDate endDate) {
-    }
-
-    /**
-     * 그룹 분담 지출 통계 관련 메서드
-     */
-    @Override
-    public PageWithSummaryResponse<GroupShareExpenseResponseDTO> getGroupShareExpensesWithPaging(
-            UUID userId, UUID groupId, ExpenseSearchDTO searchDTO, Pageable pageable
-    ) {
-        User user = findUserById(userId);
-        Group group = findGroupById(groupId);
-
-        ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
-        // 1. 필터링이 적용된 페이징 조회
-        Page<Expense> shareExpensesPage = expenseRepository.findShareExpensesByGroupAndMonthWithPagingFiltered(
-                user, group, searchDTO.year(), searchDTO.month(),
-                categoryEnum, searchDTO.search(), pageable
-        );
-
-        // 2. DTO 변환 (N+1 해결)
-        List<GroupShareExpenseResponseDTO> pageContent = convertShareExpensesToDTOs(shareExpensesPage.getContent());
-        Page<GroupShareExpenseResponseDTO> shareExpenseResponsePage = new PageImpl<>(pageContent, pageable, shareExpensesPage.getTotalElements());
-
-
-        // 3. 통계 계산을 위한 전체 데이터 조회
-        if (!shareExpenseResponsePage.hasContent()) {
-            return PageWithSummaryResponse.of(shareExpenseResponsePage, ExpenseSummary.empty());
-        }
-
-        // 전체 데이터 조회 (페이징 없이, 필터링 적용)
-        List<Expense> allShareExpenses = expenseRepository.findShareExpensesByGroupAndMonthWithPagingFiltered(
-                user, group, searchDTO.year(), searchDTO.month(),
-                categoryEnum, searchDTO.search(), Pageable.unpaged()
-        ).getContent();
-
-        List<GroupShareExpenseResponseDTO> allShareExpensesDTOs = convertShareExpensesToDTOs(allShareExpenses);
-
-        // 4. 통계 정보 계산
-        ExpenseSummary summary = calculateShareExpensesSummary(allShareExpensesDTOs);
-
-        return PageWithSummaryResponse.of(shareExpenseResponsePage, summary);
-    }
-
-    /**
-     * Expense 리스트를 GroupShareExpenseResponseDTO 리스트로 변환 (N+1 문제 해결)
-     */
-    private List<GroupShareExpenseResponseDTO> convertShareExpensesToDTOs(List<Expense> shareExpenses) {
-        if (shareExpenses.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        // 1. 모든 원본 지출 객체를 수집
-        List<Expense> originalExpenses = shareExpenses.stream()
-                .map(Expense::getOriginalExpense)
-                .filter(Objects::nonNull)
-                .distinct() // 중복 제거
-                .toList();
-
-        // 2. 모든 원본 지출의 분담 내역을 한 번에 조회
-        final Map<Long, List<SplitDataDTO>> splitsByOriginalExpenseId;
-        if (!originalExpenses.isEmpty()) {
-            splitsByOriginalExpenseId = expenseSplitRepository.findByExpenseIn(originalExpenses).stream()
-                    .collect(Collectors.groupingBy(
-                            split -> split.getExpense().getExpenseId(),
-                            Collectors.mapping(
-                                    split -> new SplitDataDTO(split.getUser().getUserId(), split.getAmount()),
-                                    Collectors.toList()
-                            )
-                    ));
-        } else {
-            splitsByOriginalExpenseId = Collections.emptyMap();
-        }
-
-        // 3. DTO 변환
-        return shareExpenses.stream()
-                .map(shareExpense -> {
-                    Expense originalExpense = shareExpense.getOriginalExpense();
-                    String groupName = shareExpense.getGroup() != null ? shareExpense.getGroup().getGroupName() : "";
-                    List<SplitDataDTO> splitData = (originalExpense != null) ?
-                            splitsByOriginalExpenseId.getOrDefault(originalExpense.getExpenseId(), Collections.emptyList()) :
-                            Collections.emptyList();
-
-                    return GroupShareExpenseResponseDTO.from(
-                            shareExpense,
-                            originalExpense,
-                            groupName,
-                            splitData
-                    );
-                })
-                .toList();
-    }
-    /**
-     * GroupShareExpenseResponseDTO 리스트로 분담 지출 통계 정보 계산
-     */
-    private ExpenseSummary calculateShareExpensesSummary(List<GroupShareExpenseResponseDTO> expenses) {
-        if (expenses.isEmpty()) {
-            return ExpenseSummary.empty();
-        }
-
-        // 전체 지출 금액 (원본 지출 기준)
-        BigDecimal totalAmount = expenses.stream()
-                .map(GroupShareExpenseResponseDTO::amount)  // 원본 지출 총액
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-        // 내 분담금 총합 (새로 추가)
-        BigDecimal myTotalShareAmount = expenses.stream()
-                .map(GroupShareExpenseResponseDTO::myShareAmount)
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-        BigDecimal averageAmount = totalAmount.divide(
-                BigDecimal.valueOf(expenses.size()), 2, RoundingMode.HALF_UP
-        );
-
-        // 최대 지출 찾기 추가
-        GroupShareExpenseResponseDTO maxExpense = expenses.stream()
-                .max(Comparator.comparing(GroupShareExpenseResponseDTO::amount))
-                .orElse(null);
-
-        long totalCount = expenses.size();
-        long myShareCount = expenses.size();  // 내 분담금 건수 (현재는 동일)
-
-        // 카테고리별 통계 계산 (내 분담금 기준)
-        Map<String, CategoryStats> categoryStats = calculateCategoryStats(
-                expenses.stream().map(e -> new ExpenseStatItem(e.category().name(), e.myShareAmount())).toList(),
-                myTotalShareAmount  // 내 분담금 총합 기준으로 변경
-        );
-
-        return new ExpenseSummary(
-                totalAmount,
-                averageAmount,
-                maxExpense.amount(),
-                totalCount,
-                myTotalShareAmount,
-                myShareCount,
-                categoryStats,
-                maxExpense.expenseId(),
-                maxExpense.title()
-        );
-    }
-
-    @Override
-    public PageWithSummaryResponse<CombinedExpenseResponseDTO> getCombinedExpenses(UUID userId, ExpenseSearchDTO searchDTO, Pageable pageable) {
-
-        User user = findUserById(userId);
-
-        // 1. 개인 지출과 분담 지출 DTO 리스트를 각각 생성
-        List<CombinedExpenseResponseDTO> personalExpenseDTOs = fetchPersonalExpenseDTOs(user, searchDTO);
-        List<CombinedExpenseResponseDTO> shareExpenseDTOs = fetchShareExpenseDTOs(user, searchDTO);
-
-        // 2. 두 리스트를 통합하고 정렬
-        List<CombinedExpenseResponseDTO> allExpenses = new ArrayList<>();
-        allExpenses.addAll(personalExpenseDTOs);
-        allExpenses.addAll(shareExpenseDTOs);
-        allExpenses.sort(Comparator.comparing(CombinedExpenseResponseDTO::date).reversed());
-
-        // 3. 통계 정보 계산
-        ExpenseSummary summary = calculateExpenseSummary(allExpenses);
-
-        // 4. 수동 페이징 처리
-        Page<CombinedExpenseResponseDTO> page = applyManualPaging(allExpenses, pageable);
-
-        return PageWithSummaryResponse.of(page, summary);
-    }
-
-    /**
-     * 사용자의 개인 지출 내역을 조회하여 DTO 리스트 변환
-     */
-    private List<CombinedExpenseResponseDTO> fetchPersonalExpenseDTOs(User user, ExpenseSearchDTO searchDTO) {
-
-        ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
-        List<Expense> personalExpenses = expenseRepository.findPersonalExpensesForCombined(
-                user, searchDTO.year(), searchDTO.month(), categoryEnum, searchDTO.search(), Pageable.unpaged()
-        ).getContent();
-
-        return personalExpenses.stream()
-                .map(CombinedExpenseResponseDTO::fromPersonal)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * 사용자의 그룹 분담 지출 내역을 조회하여 DTO 리스트 변환 (N+1 해결)
-     */
-    private List<CombinedExpenseResponseDTO> fetchShareExpenseDTOs(User user, ExpenseSearchDTO searchDTO) {
-        ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
-        List<Expense> shareExpenses = expenseRepository.findShareExpensesForCombined(
-                user, searchDTO.year(), searchDTO.month(), categoryEnum, searchDTO.search(), Pageable.unpaged()
-        ).getContent();
-
-        if (shareExpenses.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        return shareExpenses.stream()
-                .map(shareExpense -> {
-                    Expense originalExpense = shareExpense.getOriginalExpense();
-                    String groupName = shareExpense.getGroup() != null ? shareExpense.getGroup().getGroupName() : "";
-                    return CombinedExpenseResponseDTO.fromGroupShare(shareExpense, originalExpense, groupName);
-                })
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * 리스트에 대해 수동으로 페이징 처리 적용
-     */
-    private <T> Page<T> applyManualPaging(List<T> sourceList, Pageable pageable) {
-        int start = (int) pageable.getOffset();
-        if (start > sourceList.size()) {
-            return new PageImpl<>(Collections.emptyList(), pageable, sourceList.size());
-        }
-        int end = Math.min(start + pageable.getPageSize(), sourceList.size());
-        List<T> pageContent = sourceList.subList(start, end);
-        return new PageImpl<>(pageContent, pageable, sourceList.size());
-    }
-
-    /**
-     * 사용자가 해당 그룹의 멤버인지 검증
-     */
-    private void validateGroupMembership(User user, Group group) {
-        if (!groupMemberRepository.existsByGroupAndUser(group, user)) {
-            throw new CustomException(ErrorCode.GROUP_MEMBER_NOT_FOUND);
-        }
-    }
-
-    /**
-     * 카테고리 문자열을 ExpenseCategory Enum 변환
-     */
-    private ExpenseCategory parseCategory(String category) {
-        if (category == null || category.trim().isEmpty()) {
-            return null;
-        }
-        try {
-            return ExpenseCategory.valueOf(category.toUpperCase());
-        } catch (IllegalArgumentException e) {
-            // 또는 로깅 후 null 반환 등 정책에 맞게 처리
-            throw new CustomException(ErrorCode.INVALID_CATEGORY);
-        }
-    }
-
-    /**
-     * 조건에 따라 그룹 지출 내역을 데이터베이스에서 페이징하여 조회
-     */
-    private Page<Expense> fetchGroupExpenses(Group group, ExpenseSearchDTO searchDTO, Pageable pageable) {
-        String search = searchDTO.search();
-        ExpenseCategory categoryEnum = parseCategory(searchDTO.category());
-        if (search != null && !search.trim().isEmpty()) {
-            return expenseRepository.findGroupExpensesBySearchAndMonthWithPaging(
-                    group, search.trim(), searchDTO.year(), searchDTO.month(), pageable
-            );
-        } else if (categoryEnum != null) {
-            return expenseRepository.findGroupExpensesByCategoryAndMonthWithPaging(
-                    group, categoryEnum, searchDTO.year(), searchDTO.month(), pageable
-            );
-        } else {
-            return expenseRepository.findGroupExpensesByMonthWithPaging(
-                    group, searchDTO.year(), searchDTO.month(), pageable
-            );
-        }
-    }
-
-    /**
-     * CombinedExpenseResponseDTO 리스트로 통계 정보 계산
-     */
-    private ExpenseSummary calculateExpenseSummary(List<CombinedExpenseResponseDTO> expenses) {
-        if (expenses.isEmpty()) {
-            return ExpenseSummary.empty();
-        }
-
-        // 기본 통계 계산
-        BigDecimal totalAmount = expenses.stream()
-                .map(CombinedExpenseResponseDTO::myShareAmount)
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-        BigDecimal averageAmount = totalAmount.divide(
-                BigDecimal.valueOf(expenses.size()), 2, RoundingMode.HALF_UP
-        );
-
-        // 최대 지출 찾기 추가
-        CombinedExpenseResponseDTO maxExpense = expenses.stream()
-                .max(Comparator.comparing(CombinedExpenseResponseDTO::myShareAmount))
-                .orElse(null);
-
-        long totalCount = expenses.size();
-
-
-        // 카테고리별 통계 계산
-        Map<String, CategoryStats> categoryStats = calculateCategoryStats(
-                expenses.stream().map(e -> new ExpenseStatItem(e.category().name(), e.myShareAmount())).toList(),
-                totalAmount
-        );
-
-        // 개인/통합 지출의 경우 totalAmount와 myTotalShareAmount가 동일
-        return new ExpenseSummary(
-                totalAmount,
-                averageAmount,
-                maxExpense.myShareAmount(),
-                totalCount,
-                totalAmount,
-                totalCount,
-                categoryStats,
-                maxExpense.expenseId(),
-                maxExpense.title()
-        );
-    }
-
-    /**
-     * 공통 카테고리별 통계 계산
-     */
-    private Map<String, CategoryStats> calculateCategoryStats(List<ExpenseStatItem> items, BigDecimal totalAmount) {
-        Map<String, List<ExpenseStatItem>> categoryGroups = items.stream()
-                .collect(Collectors.groupingBy(ExpenseStatItem::category));
-
-        return categoryGroups.entrySet().stream()
-                .collect(Collectors.toMap(
-                        Map.Entry::getKey,
-                        entry -> {
-                            List<ExpenseStatItem> categoryExpenses = entry.getValue();
-
-                            BigDecimal categoryAmount = categoryExpenses.stream()
-                                    .map(ExpenseStatItem::amount)
-                                    .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-                            long categoryCount = categoryExpenses.size();
-
-                            BigDecimal percentage = totalAmount.compareTo(BigDecimal.ZERO) == 0
-                                    ? BigDecimal.ZERO
-                                    : categoryAmount.divide(totalAmount, 4, RoundingMode.HALF_UP)
-                                    .multiply(BigDecimal.valueOf(100));
-
-                            return new CategoryStats(categoryAmount, categoryCount, percentage);
-                        }
-                ));
-    }
-
-    /**
-     * 통계 계산용 내부 record
-     */
-    private record ExpenseStatItem(String category, BigDecimal amount) {
-    }
-
-    @Override
-    public PageWithSummaryResponse<ExpenseResponseDTO> getGroupExpensesWithStats(
-            UUID userId, UUID groupId, ExpenseSearchDTO searchDTO, Pageable pageable
-    ) {
-        User user = findUserById(userId);
-        Group group = findGroupById(groupId);
-
-        // 1. 사전 조건 검증
-        validateGroupMembership(user, group);
-
-        // 2. 페이징된 데이터 조회 (기존 fetchGroupExpenses 메서드 활용)
-        Page<Expense> groupExpensesPage = fetchGroupExpenses(group, searchDTO, pageable);
-
-        if (!groupExpensesPage.hasContent()) {
-            return PageWithSummaryResponse.of(Page.empty(pageable), ExpenseSummary.empty());
-        }
-
-        // 3. Page<Expense>를 Page<ExpenseResponseDTO>로 변환
-        List<ExpenseResponseDTO> pageContent = convertToExpenseResponseDTOs(groupExpensesPage.getContent());
-        Page<ExpenseResponseDTO> expenseResponsePage = new PageImpl<>(pageContent, pageable, groupExpensesPage.getTotalElements());
-
-
-        // 4. 통계 계산을 위한 전체 데이터 조회 (기존 fetchGroupExpenses 활용)
-        List<Expense> allExpensesForStats = fetchGroupExpenses(group, searchDTO, Pageable.unpaged()).getContent();
-        List<ExpenseResponseDTO> allExpenseDTOsForStats = convertToExpenseResponseDTOs(allExpensesForStats);
-
-
-        // 5. 통계 정보 계산
-        ExpenseSummary summary = calculateGroupExpensesSummary(allExpenseDTOsForStats);
-
-        return PageWithSummaryResponse.of(expenseResponsePage, summary);
-    }
-
-    /**
-     * 지출 목록을 DTO로 변환
-     */
-    private List<ExpenseResponseDTO> convertToExpenseResponseDTOs(List<Expense> expenses) {
-        return expenses.stream()
-                .map(expense -> {
-                    List<SplitDataDTO> splitData = getSplitData(expense);
-                    return ExpenseResponseDTO.from(expense, splitData);
-                })
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * ExpenseResponseDTO 리스트로 그룹 지출 통계 정보 계산
-     */
-    private ExpenseSummary calculateGroupExpensesSummary(List<ExpenseResponseDTO> expenses) {
-        if (expenses.isEmpty()) {
-            return ExpenseSummary.empty();
-        }
-
-        BigDecimal totalAmount = expenses.stream()
-                .map(ExpenseResponseDTO::amount)
-                .reduce(BigDecimal.ZERO, BigDecimal::add);
-
-        BigDecimal averageAmount = totalAmount.divide(
-                BigDecimal.valueOf(expenses.size()), 2, RoundingMode.HALF_UP
-        );
-
-        // 최대 지출 찾기 추가
-        ExpenseResponseDTO maxExpense = expenses.stream()
-                .max(Comparator.comparing(ExpenseResponseDTO::amount))
-                .orElse(null);
-
-        long totalCount = expenses.size();
-
-        // 기존 calculateCategoryStats 메서드 활용
-        Map<String, CategoryStats> categoryStats = calculateCategoryStats(
-                expenses.stream().map(e -> new ExpenseStatItem(e.category().name(), e.amount())).toList(),
-                totalAmount
-        );
-
-        return new ExpenseSummary(
-                totalAmount,
-                averageAmount,
-                maxExpense.amount(),
-                totalCount,
-                totalAmount,
-                totalCount,
-                categoryStats,
-                maxExpense.expenseId(),
-                maxExpense.title()
-        );
-    }
 
     /**
      * LLM 지출 분석
-     * @param userId 사용자의 ID
+     *
+     * @param userId     사용자의 ID
      * @param requestDTO 시작일, 종료일
      * @return 4가지 유형 - 예산 사용률, 일 평균 지출(월말 예상), 분석 결과, 카테고리별 상세 분석
      */
     @Override
-    @Transactional
     public AnalyzeExpenseResponseDTO analyzeExpenses(UUID userId, AnalyzeExpenseRequestDTO requestDTO) {
         User user = userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
 
@@ -1055,9 +400,9 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
         BigDecimal totalSpending = calculateTotalSpending(expenses);
 
         AnalyzeExpenseResponseDTO.BudgetUsage budgetUsage = calculateBudgetUsage(totalSpending, totalBudget);
-        AnalyzeExpenseResponseDTO.DailySpending dailySpending = calculateDailySpending(totalSpending,requestDTO.startDate(),requestDTO.endDate());
+        AnalyzeExpenseResponseDTO.DailySpending dailySpending = calculateDailySpending(totalSpending, requestDTO.startDate(), requestDTO.endDate());
 
-        List<AnalyzeExpenseResponseDTO.CategoryDetail> categoryDetails = calculateCategoryDetails(expenses,totalSpending);
+        List<AnalyzeExpenseResponseDTO.CategoryDetail> categoryDetails = calculateCategoryDetails(expenses, totalSpending);
 
         AnalyzeExpenseResponseDTO.Suggestion suggestion = getLlmAnalysisResult(expenses);
 
@@ -1065,26 +410,9 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
 
         AnalyzeExpenseResponseDTO.AnalysisResult analysisResult = new AnalyzeExpenseResponseDTO.AnalysisResult(mainFinding, suggestion);
 
-        ExpenseAnalysisHistory savedHistory = saveAnalysisHistory(user,requestDTO,budgetUsage,dailySpending,analysisResult);
+        ExpenseAnalysisHistory savedHistory = saveAnalysisHistory(user, requestDTO, budgetUsage, dailySpending, analysisResult);
 
-        return new AnalyzeExpenseResponseDTO(savedHistory.getId(), budgetUsage,dailySpending,analysisResult,categoryDetails);
-    }
-
-    /**
-     * LLM 지출 분석 기록 조회
-     */
-    @Override
-    public PageWithSummaryResponse<ExpenseAnalysisHistoryDTO> getExpenseAnalysisHistory(UUID userId, Pageable pageable) {
-
-//        User user = userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
-        User user = findUserById(userId);
-
-        Page<ExpenseAnalysisHistory> historyPage = expenseAnalysisHistoryRepository.findByUser(user, pageable);
-
-        Page<ExpenseAnalysisHistoryDTO> historyDTOPage = historyPage.map(ExpenseAnalysisHistoryDTO::from);
-
-        return PageWithSummaryResponse.of(historyDTOPage, ExpenseSummary.empty());
-
+        return new AnalyzeExpenseResponseDTO(savedHistory.getId(), budgetUsage, dailySpending, analysisResult, categoryDetails);
     }
 
     /**
@@ -1115,7 +443,6 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
      * LLM 지출 분석 피드백
      */
     @Override
-    @Transactional
     public String toggleFeedback(Long historyId, UUID userid, String type) {
         ExpenseAnalysisHistory history = expenseAnalysisHistoryRepository.findById(historyId)
                 .orElseThrow(() -> new CustomException(ErrorCode.HISTORY_NOT_FOUND));
@@ -1127,30 +454,32 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
         boolean isUp = "up".equals(type);
         boolean isDown = "down".equals(type);
 
-        if(!isUp && !isDown) {
+        if (!isUp && !isDown) {
             throw new CustomException(ErrorCode.INVALID_HISTORY_REQUEST);
         }
         // 현재 상태와 같은 버튼을 다시 누르면 피드백 취소
-        if((isUp && Boolean.TRUE.equals(history.getUp())) || (isDown && Boolean.TRUE.equals(history.getDown()))){
-            history.feedback(null,null);
+        if ((isUp && Boolean.TRUE.equals(history.getUp())) || (isDown && Boolean.TRUE.equals(history.getDown()))) {
+            history.feedback(null, null);
             return "CANCELED";
-        } else{
+        } else {
             // 새로운 피드백 설정
-            history.feedback(isUp,isDown);
+            history.feedback(isUp, isDown);
             return "APPLIED";
         }
     }
 
     /**
      * 사용자의 예산 정보를 조회합니다.
+     *
      * @param userId 사용자의 UUID
      * @return 사용자의 예산
      */
-    private BigDecimal getUserBudget(UUID userId){
+    private BigDecimal getUserBudget(UUID userId) {
         User user = userRepository.findById(userId).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
 
         return BigDecimal.valueOf(user.getUserBudget());
     }
+
     /**
      * 지출 내역 리스트를 받아 총 지출액 계산
      */
@@ -1159,38 +488,39 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
                 .map(Expense::getAmount)
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
+
     /**
-    * 예산 사용률 객체 생성
-    */
-    private AnalyzeExpenseResponseDTO.BudgetUsage calculateBudgetUsage(BigDecimal totalSpending, BigDecimal totalBudget){
+     * 예산 사용률 객체 생성
+     */
+    private AnalyzeExpenseResponseDTO.BudgetUsage calculateBudgetUsage(BigDecimal totalSpending, BigDecimal totalBudget) {
         if (totalBudget.compareTo(BigDecimal.ZERO) == 0) {
             return new AnalyzeExpenseResponseDTO.BudgetUsage(0.0, totalSpending, totalBudget);
         }
-        BigDecimal percentage = totalSpending.multiply(BigDecimal.valueOf(100)).divide(totalBudget,2,RoundingMode.HALF_UP);
+        BigDecimal percentage = totalSpending.multiply(BigDecimal.valueOf(100)).divide(totalBudget, 2, RoundingMode.HALF_UP);
         return new AnalyzeExpenseResponseDTO.BudgetUsage(percentage.doubleValue(), totalSpending, totalBudget);
     }
 
     /**
      * 일평균, 월말 지출 객체 생성
      */
-    private AnalyzeExpenseResponseDTO.DailySpending calculateDailySpending(BigDecimal totalSpending, LocalDate startDate, LocalDate endDate){
+    private AnalyzeExpenseResponseDTO.DailySpending calculateDailySpending(BigDecimal totalSpending, LocalDate startDate, LocalDate endDate) {
         long days = java.time.temporal.ChronoUnit.DAYS.between(startDate, LocalDate.now()) + 1;
-        BigDecimal averageSoFar = totalSpending.divide(BigDecimal.valueOf(days),0,RoundingMode.HALF_UP);
+        BigDecimal averageSoFar = totalSpending.divide(BigDecimal.valueOf(days), 0, RoundingMode.HALF_UP);
         int daysInMonth = endDate.lengthOfMonth();
         BigDecimal estimatedEom = averageSoFar.multiply(BigDecimal.valueOf(daysInMonth)); // 월말 예상 지출
 
-        return new AnalyzeExpenseResponseDTO.DailySpending(averageSoFar,estimatedEom);
+        return new AnalyzeExpenseResponseDTO.DailySpending(averageSoFar, estimatedEom);
     }
 
     /**
      * 카테고리별 상세 분석 객체 리스트 생성
      */
-    private List<AnalyzeExpenseResponseDTO.CategoryDetail> calculateCategoryDetails(List<Expense> expenses, BigDecimal totalSpending){
-        if(totalSpending.compareTo(BigDecimal.ZERO) == 0){
+    private List<AnalyzeExpenseResponseDTO.CategoryDetail> calculateCategoryDetails(List<Expense> expenses, BigDecimal totalSpending) {
+        if (totalSpending.compareTo(BigDecimal.ZERO) == 0) {
             return Collections.emptyList();
         }
 
-        Map<ExpenseCategory,List<Expense>> expensesByCategory = expenses.stream()
+        Map<ExpenseCategory, List<Expense>> expensesByCategory = expenses.stream()
                 .collect(Collectors.groupingBy(Expense::getCategory));
 
         return expensesByCategory.entrySet().stream()
@@ -1198,17 +528,17 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
                     ExpenseCategory category = entry.getKey();
                     List<Expense> categoryExpenses = entry.getValue();
                     BigDecimal categoryTotal = calculateTotalSpending(categoryExpenses);
-                    double percentage = categoryTotal.divide(totalSpending,4,RoundingMode.HALF_UP)
+                    double percentage = categoryTotal.divide(totalSpending, 4, RoundingMode.HALF_UP)
                             .multiply(BigDecimal.valueOf(100)).doubleValue();
                     int count = categoryExpenses.size();
-                    return new AnalyzeExpenseResponseDTO.CategoryDetail(category.getDescription(),percentage,categoryTotal,count);
+                    return new AnalyzeExpenseResponseDTO.CategoryDetail(category.getDescription(), percentage, categoryTotal, count);
                 }).sorted(Comparator.comparing(AnalyzeExpenseResponseDTO.CategoryDetail::percentage).reversed()).toList();
     }
 
     /**
      * LLM으로부터 개선 제안을 받아옴
      */
-    private AnalyzeExpenseResponseDTO.Suggestion getLlmAnalysisResult(List<Expense> expenses){
+    private AnalyzeExpenseResponseDTO.Suggestion getLlmAnalysisResult(List<Expense> expenses) {
         // LLM에 전달할 데이터만 동적으로 가공
         List<Map<String, Object>> llmExpenseData = expenses.stream()
                 .map(expense -> {
@@ -1225,10 +555,10 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
                     return data;
                 })
                 .toList();
-        try{
+        try {
             String expenseJson = objectMapper.writeValueAsString(llmExpenseData);
             return expenseAnalyzer.analyzerExpenseData(expenseJson);
-        } catch (JsonProcessingException e){
+        } catch (JsonProcessingException e) {
             log.error("LLM 전송용 DTO to JSON 변환 실패");
             return new AnalyzeExpenseResponseDTO.Suggestion("데이터 처리중 오류 발생", "잠시 후 다시 시도해주세요.", "오류", "오류");
         }
@@ -1242,15 +572,16 @@ public class ExpenseV2ServiceImpl implements ExpenseService {
         AnalyzeExpenseResponseDTO.DailySpending dailySpending = new AnalyzeExpenseResponseDTO.DailySpending(BigDecimal.ZERO, BigDecimal.ZERO);
 
         AnalyzeExpenseResponseDTO.AnalysisResult analysisResult = new AnalyzeExpenseResponseDTO.AnalysisResult("분석할 지출 내역이 없습니다.",
-                new AnalyzeExpenseResponseDTO.Suggestion("지출 내역을 추가하고 다시 시도해주세요.","없음","없음", "없음"));
+                new AnalyzeExpenseResponseDTO.Suggestion("지출 내역을 추가하고 다시 시도해주세요.", "없음", "없음", "없음"));
 
-        return new AnalyzeExpenseResponseDTO(null,budgetUsage,dailySpending,analysisResult,Collections.emptyList());
+        return new AnalyzeExpenseResponseDTO(null, budgetUsage, dailySpending, analysisResult, Collections.emptyList());
     }
+
     /**
      * 가장 지출이 큰 카테고리를 찾아 분석요약(mainFinding) 생성
      */
     private String createMainFinding(List<AnalyzeExpenseResponseDTO.CategoryDetail> categoryDetails) {
-        if (categoryDetails == null || categoryDetails.isEmpty()){
+        if (categoryDetails == null || categoryDetails.isEmpty()) {
             return "주요 지출 항목이 없습니다.";
         }
         AnalyzeExpenseResponseDTO.CategoryDetail maxCategoryDetail = categoryDetails.get(0);
